+++ bocadillo/routing/checks.py
@@ -1,11 +1,11 @@
-"""Start-up check and validation utilities."""
+"""Routing check and validation utilities."""
 import inspect
 from string import Formatter
 from typing import List
 
-from bocadillo.constants import ALL_HTTP_METHODS
-from bocadillo.view import get_declared_method_views, View, get_view_name
 from .exceptions import RouteDeclarationError
+from ..constants import ALL_HTTP_METHODS
+from ..view import get_declared_method_views, View, get_view_name
 
 
 def check_route(pattern: str, view: View, methods: List[str]) -> None:
